# temp_backup/specs/001-ai-robotics-platform/contracts/api.yaml
openapi: 3.0.3
info:
  title: Physical AI & Humanoid Robotics Textbook API
  description: API for the interactive textbook platform, handling chat and content personalization.
  version: 1.0.0
servers:
  - url: /api

paths:
  /chat:
    post:
      summary: RAG Agent Interaction
      description: Receives a user's question and returns an answer sourced from the textbook content using a RAG agent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
                  description: The user's question.
                  example: "How do I set up a ROS 2 workspace?"
                session_id:
                  type: string
                  format: uuid
                  description: The user's session ID for conversation history.
      responses:
        '200':
          description: A successful response.
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                    description: The answer generated by the RAG agent.
                    example: "To set up a ROS 2 workspace, you first need to create a directory, then..."
                  sources:
                    type: array
                    items:
                      type: object
                      properties:
                        chapter_id:
                          type: string
                          example: "ros-2-setup"
                        similarity_score:
                          type: number
                          format: float
                          example: 0.92
        '400':
          description: Bad request, e.g., missing question.

  /personalize:
    post:
      summary: Personalize Content
      description: Receives the generic content of a chapter and returns a version rewritten and adapted to the user's hardware background.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  description: The raw Markdown content of the chapter to be personalized.
                  example: "## Installation\nTo install the necessary tools, run..."
                user_id:
                  type: string
                  format: uuid
                  description: The ID of the user requesting personalization, used to retrieve their hardware background.
      responses:
        '200':
          description: A successful response.
          content:
            application/json:
              schema:
                type: object
                properties:
                  personalized_content:
                    type: string
                    description: The rewritten Markdown content adapted for the user.
                    example: "## Installation (for RTX4090)\nTo install the necessary tools with CUDA support, run..."
        '404':
          description: User not found.